[dotfiles] function: new function to open a file at a specific line thanks to sack
Use the 'sack' bash script and read the result in the $SACH_SHORTCUT file in
order to open the file at a specific file after searching a pattern in subdirs.

Sack is a wrapper for grep, ack or ag in order to save the 'lineno filename'
in a dedicated file and add some extra info in the {grep, ack, ag} output in
order to quickly open a file a a specific line.